Refactor ChunkGenerator
