+++ RealWorld/generation/CMakeLists.txt
@@ -2,8 +2,6 @@
 add_subdirectory(shaders)
 
 target_sources(RealWorld PRIVATE 
-    ChunkGenerator.hpp          ChunkGenerator.cpp
-    GenerationPC.hpp            
-    TerrainGenerator.hpp        TerrainGenerator.cpp
-    TreeGenerator.hpp           TreeGenerator.cpp
+    ChunkGenerator.hpp          ChunkGeneratorMain.cpp
+    ChunkGeneratorTerrain.cpp   ChunkGeneratorTree.cpp
 )
