+++ RealWorld/generation/ChunkGenerator.hpp
@@ -2,14 +2,21 @@
  *  @author    Dubsky Tomas
  */
 #pragma once
-#include <RealWorld/generation/TerrainGenerator.hpp>
-#include <RealWorld/generation/TreeGenerator.hpp>
+#include <RealEngine/graphics/CommandBuffer.hpp>
+#include <RealEngine/graphics/descriptors/DescriptorSet.hpp>
+#include <RealEngine/graphics/pipelines/Pipeline.hpp>
+#include <RealEngine/graphics/textures/Texture.hpp>
+
+#include <RealWorld/constants/generation.hpp>
+#include <RealWorld/generation/shaders/AllShaders.hpp>
 
 namespace rw {
 
 /**
  * @brief Generates new chunks
- * @details The logic is divided into more classes: TerrainGenerator and TreeGenerator
+ * @details The class is implemented in multiple files:
+ *          ChunkGeneratorMain.cpp, ChunkGeneratorTerrain.cpp and
+ *          ChunkGeneratorTree.cpp.
  */
 class ChunkGenerator {
 public:
@@ -38,9 +45,49 @@ public:
     );
 
 protected:
-    GenerationPC     m_genPC;
-    TerrainGenerator m_terrainGen;
-    TreeGenerator    m_treeGen;
+    void prepareToGenerate(const vk::CommandBuffer& commandBuffer);
+    void generateBasicTerrain(const vk::CommandBuffer& commandBuffer);
+    void consolidateEdges(const vk::CommandBuffer& commandBuffer);
+    void selectVariant(const vk::CommandBuffer& commandBuffer);
+    void finishGeneration(
+        const vk::CommandBuffer& commandBuffer,
+        const re::Texture&       dstTex,
+        const glm::ivec2&        dstOffset
+    );
+
+    vk::ImageMemoryBarrier2 stepBarrier() const; /**< Helper func */
+
+    struct GenerationPC {
+        glm::ivec2 chunkOffsetTi;
+        int        seed;
+        glm::uint  storeLayer;
+        glm::uint  edgeConsolidationPromote;
+        glm::uint  edgeConsolidationReduce;
+    };
+
+    GenerationPC m_genPC;
+
+    re::PipelineLayout m_pipelineLayout{{}, {.comp = generateStructure_comp}};
+    re::Pipeline       m_generateStructurePl{
+              {.pipelineLayout = *m_pipelineLayout}, {.comp = generateStructure_comp}};
+    re::Pipeline m_consolidateEdgesPl{
+        {.pipelineLayout = *m_pipelineLayout}, {.comp = consolidateEdges_comp}};
+    re::Pipeline m_selectVariantPl{
+        {.pipelineLayout = *m_pipelineLayout}, {.comp = selectVariant_comp}};
+    re::DescriptorSet m_descSet{m_pipelineLayout, 0u};
+
+    re::Texture m_tilesTex{re::TextureCreateInfo{
+        .format = vk::Format::eR8G8B8A8Uint,
+        .extent = {k_genChunkSize.x, k_genChunkSize.y, 1u},
+        .layers = 2,
+        .usage  = vk::ImageUsageFlagBits::eStorage |
+                 vk::ImageUsageFlagBits::eTransferSrc,
+        .initialLayout = vk::ImageLayout::eGeneral}};
+    re::Texture m_materialTex{re::TextureCreateInfo{
+        .format        = vk::Format::eR8G8B8A8Uint,
+        .extent        = {k_genChunkSize.x, k_genChunkSize.y, 1u},
+        .usage         = vk::ImageUsageFlagBits::eStorage,
+        .initialLayout = vk::ImageLayout::eGeneral}};
 };
 
 } // namespace rw
