+++ RealWorld/generation/ChunkGeneratorTree.cpp
@@ -1,0 +1,8 @@
+﻿/*!
+ *  @author    Dubsky Tomas
+ */
+#include <RealWorld/generation/ChunkGenerator.hpp>
+
+namespace rw {
+
+} // namespace rw
