--- RealWorld/generation/TreeGenerator.cpp
@@ -1,12 +1,0 @@
-﻿/*!
- *  @author    Dubsky Tomas
- */
-#include <RealWorld/generation/TreeGenerator.hpp>
-
-namespace rw {
-
-TreeGenerator::TreeGenerator(GenerationPC& genPC)
-    : m_genPC(genPC) {
-}
-
-} // namespace rw
